+++ argilla-frontend/translation/en.js
@@ -6,6 +6,7 @@ export default {
   dynamic_label_selection: "Label (dynamic)",
   table_form_placeholder: "Copy and paste a table you'd like to edit here",
   span: "Span",
+  table: "Table",
   text: "Text",
   image: "Image",
   rating: "Rating",
@@ -337,6 +338,7 @@ export default {
       ranking: "Ranking",
       multi_label_selection: "Multi-label",
       span: "Span",
+      table: "Table",
       "no mapping": "No mapping",
     },
     questionId: {
@@ -346,6 +348,7 @@ export default {
       ranking: "ranking",
       multi_label_selection: "multi-label",
       span: "span",
+      table: "table",
     },
   },
   persistentStorage: {