+++ argilla-server/src/argilla_server/api/schemas/v1/questions.py
@@ -274,6 +274,21 @@ class SpanQuestionSettingsUpdate(UpdateSchema):
     allow_overlapping: Optional[bool]
 
 
+# Table question
+class TableQuestionSettings(BaseModel):
+    type: Literal[QuestionType.table]
+
+
+class TableQuestionSettingsCreate(UniqueValuesCheckerMixin):
+    type: Literal[QuestionType.table]
+
+
+class TableQuestionSettingsUpdate(UpdateSchema):
+    type: Literal[QuestionType.table]
+
+    __non_nullable_fields__ = {}
+
+
 QuestionSettings = Annotated[
     Union[
         TextQuestionSettings,
@@ -282,6 +297,7 @@ QuestionSettings = Annotated[
         MultiLabelSelectionQuestionSettings,
         RankingQuestionSettings,
         SpanQuestionSettings,
+        TableQuestionSettings,
     ],
     Field(..., discriminator="type"),
 ]
@@ -322,6 +338,7 @@ QuestionSettingsCreate = Annotated[
         MultiLabelSelectionQuestionSettingsCreate,
         RankingQuestionSettingsCreate,
         SpanQuestionSettingsCreate,
+        TableQuestionSettingsCreate,
     ],
     Field(discriminator="type"),
 ]
@@ -335,6 +352,7 @@ QuestionSettingsUpdate = Annotated[
         MultiLabelSelectionQuestionSettingsUpdate,
         RankingQuestionSettingsUpdate,
         SpanQuestionSettingsUpdate,
+        TableQuestionSettingsUpdate,
     ],
     Field(..., discriminator="type"),
 ]