+++ argilla/src/argilla/_models/__init__.py
@@ -50,6 +50,8 @@ from argilla._models._settings._questions import (
     RatingQuestionModel,
     SpanQuestionModel,
     SpanQuestionSettings,
+    TableQuestionModel,
+    TableQuestionSettings,
     TextQuestionModel,
     TextQuestionSettings,
 )